requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
cryptography==42.0.5
python-dotenv==1.0.1
